requirements.txt
fastapi
uvicorn
asyncpg
pyjwt
grpcio
grpcio-tools
pydantic
